Rename class ControllerUtils -> PlatformUtil
