+++ library/controller/kmp-tor-controller-common/src/jsMain/kotlin/io/matthewnelson/kmp/tor/controller/common/internal/PlatformUtil.kt
@@ -18,14 +18,45 @@ package io.matthewnelson.kmp.tor.controller.common.internal
 import io.matthewnelson.kmp.tor.common.annotation.InternalTorApi
 
 // TODO: Fill out
-actual object ControllerUtils {
+actual object PlatformUtil {
     @InternalTorApi
     actual fun localhostAddress(): String = "127.0.0.1"
 
-    actual val isDarwin: Boolean = false
-    actual val isLinux: Boolean = true
-    actual val isMingw: Boolean = false
+    actual val isDarwin: Boolean by lazy {
+        platform == "darwin"
+    }
+    actual val isLinux: Boolean by lazy {
+        platform == "linux" || platform == "android"
+    }
+    actual val isMingw: Boolean by lazy {
+        platform == "win32"
+    }
 
     @InternalTorApi
-    actual val hasControlUnixDomainSocketSupport: Boolean = true
+    actual val hasControlUnixDomainSocketSupport: Boolean = false
 }
+
+/**
+ * Returns:
+ *  - aix
+ *  - android
+ *  - darwin
+ *  - freebsd
+ *  - linux
+ *  - openbsd
+ *  - sunos
+ *  - win32
+ * */
+private val platform: String
+    get() {
+        return os?.platform() as? String ?: "linux"
+    }
+
+private val os: dynamic
+    get() {
+        return try {
+            js("require('os')")
+        } catch (_: Throwable) {
+            null
+        }
+    }
